docs: README networks section
## Motivation
I'm not sure where the source of truth is for the version(s) of software that are running on our networks. This repo seems like a good candidate for such a table. I'm concerned this table won't be updated when software on our networks is updated so please comment if there's a better repo for such a table.